[Java API] Filters dealing with time now take double arguments
git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@587 b725f6ce-2418-43e8-b868-1e0ad3d2b257

+++ src/java/muscle/core/conduit/filter/BlockAfterTimeFilter.java
@@ -36,7 +36,7 @@ public class BlockAfterTimeFilter<E extends Serializable> extends AbstractObserv
 	private final static Logger logger = Logger.getLogger(BlockAfterTimeFilter.class.getName());
 
 	/** @param newMaxTime seconds after which the filter blocks */
-	public BlockAfterTimeFilter(int newMaxTime) {
+	public BlockAfterTimeFilter(double newMaxTime) {
 		super();
 		maxTime = new Timestamp(newMaxTime);
 	}
